Moving stubs into kmp module.
