--- app/src/main/java/com/adammcneilly/spacenerd/core/ui/components/ImageWrapper.kt
@@ -1,69 +1,0 @@
-package com.adammcneilly.spacenerd.core.ui.components
-
-import androidx.compose.foundation.Image
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import coil3.compose.AsyncImage
-import com.adammcneilly.spacenerd.shared.core.displaymodels.ImageDisplayModel
-
-@Composable
-fun ImageWrapper(
-    image: ImageDisplayModel,
-    contentDescription: String?,
-    modifier: Modifier = Modifier,
-    contentScale: ContentScale = ContentScale.Fit,
-) {
-    when (image) {
-        is ImageDisplayModel.AndroidLocal -> {
-            LocalImage(
-                image = image,
-                contentDescription = contentDescription,
-                contentScale = contentScale,
-                modifier = modifier,
-            )
-        }
-        is ImageDisplayModel.Remote -> {
-            RemoteImage(
-                image = image,
-                contentDescription = contentDescription,
-                contentScale = contentScale,
-                modifier = modifier,
-            )
-        }
-        else -> {
-            // No-Op
-        }
-    }
-}
-
-@Composable
-private fun RemoteImage(
-    image: ImageDisplayModel.Remote,
-    contentDescription: String?,
-    contentScale: ContentScale,
-    modifier: Modifier,
-) {
-    AsyncImage(
-        model = image.url,
-        contentDescription = contentDescription,
-        contentScale = contentScale,
-        modifier = modifier,
-    )
-}
-
-@Composable
-private fun LocalImage(
-    image: ImageDisplayModel.AndroidLocal,
-    contentDescription: String?,
-    contentScale: ContentScale,
-    modifier: Modifier,
-) {
-    Image(
-        painter = painterResource(image.drawableRes),
-        contentDescription = contentDescription,
-        contentScale = contentScale,
-        modifier = modifier,
-    )
-}
