--- app/src/main/java/com/adammcneilly/spacenerd/core/ui/utils/PaddingValues.kt
@@ -1,23 +1,0 @@
-package com.adammcneilly.spacenerd.core.ui.utils
-
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.LayoutDirection
-
-operator fun PaddingValues.plus(
-    that: PaddingValues,
-): PaddingValues {
-    return object : PaddingValues {
-        override fun calculateBottomPadding(): Dp = this@plus.calculateBottomPadding() + that.calculateBottomPadding()
-
-        override fun calculateLeftPadding(
-            layoutDirection: LayoutDirection,
-        ): Dp = this@plus.calculateLeftPadding(layoutDirection) + that.calculateLeftPadding(layoutDirection)
-
-        override fun calculateRightPadding(
-            layoutDirection: LayoutDirection,
-        ): Dp = this@plus.calculateRightPadding(layoutDirection) + that.calculateRightPadding(layoutDirection)
-
-        override fun calculateTopPadding(): Dp = this@plus.calculateTopPadding() + that.calculateTopPadding()
-    }
-}
